fix: correct `request`/`requests` typo in bidi stream
Updates the signature for `bidi_streamer` to use the keyword `request`
rather than `requests`.  This is in accordance with the signatures
specified in `GRPC::Interceptor::ClientInterceptor`. [1]

The previous use of `request` as a keyword seems to be a bug.  It would
throw an exception when any bidi method is invoked via this interceptor.

[1] https://github.com/grpc/grpc/blob/v1.66.1/src/ruby/lib/grpc/generic/interceptors.rb#L86

Resolves #1